feat: add api for updating and deleting test case

+++ src/app/api/delete/route.ts
@@ -6,13 +6,23 @@ export async function POST(request: NextRequest) {
     const jwtToken = request.cookies.get('jwtToken')
 
     try {
-        const res = await axios.delete(`${link}/${id}`, {
-            headers: {
-                Authorization: `Bearer ${jwtToken?.value}`,
-            },
-        });
-    
-        return NextResponse.json(res.data, { status: 200 });
+        if(id){
+            const res = await axios.delete(`${link}/${id}`, {
+                headers: {
+                    Authorization: `Bearer ${jwtToken?.value}`,
+                },
+            });
+        
+            return NextResponse.json(res.data, { status: 200 });
+        }
+
+        const res = await axios.delete(`${link}`, {
+                headers: {
+                    Authorization: `Bearer ${jwtToken?.value}`,
+                },
+            });
+        
+            return NextResponse.json(res.data, { status: 200 });
     } catch (error) {
         return NextResponse.json(
             { message: `Have an error: ${error}` },
